Cache disk blocks on file open to avoid indirection on read.
Cache the disk blocks into a linked list when a file is open, so that
we can avoid disk reads of indirect blocks to find the block when doing
a read. We can now just jump directly to the appropriate disk block and
cache it in memory for future reads. Depending on the size of the file
this can void 2-3 disk reads per call to read.